requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.0
pytest==7.4.0
